% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{name_to_PAO1}
\alias{name_to_PAO1}
\title{Convert from 3 or 4 letter P. aeruginosa gene name to PAO1 number syntax}
\usage{
name_to_PAO1(x)
}
\arguments{
\item{x}{A 3 or 4 letter gene name}
}
\value{
The corresponding PAO1 number
}
\description{
Convert from 3 or 4 letter P. aeruginosa gene name to PAO1 number syntax
}
\examples{
name_to_PAO1('dnaA')
}
