% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\docType{data}
\name{PA14_old_v_new}
\alias{PA14_old_v_new}
\title{Misc. annotation files for converting gene names and numbers}
\format{
An object of class \code{data.frame} with 5864 rows and 15 columns.
}
\usage{
PA14_old_v_new
}
\description{
A small set of functions to convert between gene names in Pseudomonas,
Particularly PAO1 numbers and PA14 number interconversion, but also compatible with other species/strains curated in pseudomonas.com
}
\keyword{datasets}
