% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{PA14_to_name}
\alias{PA14_to_name}
\title{Convert from P. aeruginosa gene PA14 number to 3 or 4 letter gene name}
\usage{
PA14_to_name(x)
}
\arguments{
\item{x}{A PA14 number}
}
\value{
The corresponding 3 or 4 letter gene name
}
\description{
Convert from P. aeruginosa gene PA14 number to 3 or 4 letter gene name
}
\examples{
PA14_to_name('PA14_00001')
PAO1_to_name(PA14_to_PAO1(PA14_to_name('PA14_0001')))
}
