% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{PA14_to_PA14new}
\alias{PA14_to_PA14new}
\title{Convert from P. aeruginosa gene PA14 number to Refseq UCBPP-PA14 genome gene number}
\usage{
PA14_to_PA14new(x)
}
\arguments{
\item{x}{A PA14 number}
}
\value{
The homologous PA14 number in current (2024-04) RefSeq UCBPP-PA14 genome
}
\description{
Convert from P. aeruginosa gene PA14 number to Refseq UCBPP-PA14 genome gene number
}
\examples{
PA14_to_PA14new('PA14_00001')
}
