% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{PA14_to_PAO1}
\alias{PA14_to_PAO1}
\title{Convert from P. aeruginosa gene PA14 number to PAO1 number}
\usage{
PA14_to_PAO1(x)
}
\arguments{
\item{x}{A PA14 number}
}
\value{
The homologous PAO1 number
}
\description{
Convert from P. aeruginosa gene PA14 number to PAO1 number
}
\examples{
PA14_to_PAO1('PA14_00001')
}
