% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{PA14new_to_PA14}
\alias{PA14new_to_PA14}
\title{Convert from P. aeruginosa Refseq UCBPP-PA14 genome gene number to older PA14 gene number nomenclature}
\usage{
PA14new_to_PA14(x)
}
\arguments{
\item{x}{A PA14_RS number in the current (2024-04) Refseq UCBPP-PA14 genome}
}
\value{
The homologous PA14 number in older PA14 format (used in all other functions in this package)
}
\description{
Convert from P. aeruginosa Refseq UCBPP-PA14 genome gene number to older PA14 gene number nomenclature
}
\examples{
PA14new_to_PA14('PA14_RS00001')
}
