% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\docType{data}
\name{PAO1_orth}
\alias{PAO1_orth}
\title{Misc. annotation files for converting gene names and numbers}
\format{
An object of class \code{data.frame} with 380042 rows and 8 columns.
}
\usage{
PAO1_orth
}
\description{
A small set of functions I frquently use to convert between gene names,
PAO1 numbers and PA14 numbers
}
\keyword{datasets}
