% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{PAO1_to_name}
\alias{PAO1_to_name}
\title{Convert from P. aeruginosa gene PAO1 number to 3 or 4 letter gene name}
\usage{
PAO1_to_name(x)
}
\arguments{
\item{x}{A  PAO1 number}
}
\value{
The corresponding 3 or 4 letter gene name
}
\description{
Convert from P. aeruginosa gene PAO1 number to 3 or 4 letter gene name
}
\examples{
PAO1_to_name('PA0001')
}
