% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{PAO1_to_PA14}
\alias{PAO1_to_PA14}
\title{Convert from P. aeruginosa gene PAO1 number to PA14 number}
\usage{
PAO1_to_PA14(x)
}
\arguments{
\item{x}{A PAO1 number}
}
\value{
The homologous PA14 number
}
\description{
Convert from P. aeruginosa gene PAO1 number to PA14 number
}
\examples{
PAO1_to_PA14('PA0001')
}
