% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{strain_to_PA14}
\alias{strain_to_PA14}
\title{Convert from Pseudomonas strain gene number (any supported strain) to PA14 number}
\usage{
strain_to_PA14(x, y)
}
\arguments{
\item{x}{A Pseudomonas strain's gene number}

\item{y}{A Pseudomonas strain name, among supported strains}
}
\value{
The homologous PA14 number
}
\description{
Convert from Pseudomonas strain gene number (any supported strain) to PA14 number
}
\examples{
Strain_to_PA14('PFLU0001', 'Pseudomonas fluorescens SBW25')
}
