% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{strain_to_PAO1}
\alias{strain_to_PAO1}
\title{Convert from Pseudomonas strain gene number (any supported strain) to PAO1 number}
\usage{
strain_to_PAO1(x, y)
}
\arguments{
\item{x}{A Pseudomonas strain's gene number}

\item{y}{A Pseudomonas strain name, among supported strains}
}
\value{
The homologous PAO1 number
}
\description{
Convert from Pseudomonas strain gene number (any supported strain) to PAO1 number
}
\examples{
Strain_to_PAO1('PFL_0001', 'Pseudomonas protegens Pf-5')
}
